Better search page: link to project, copy key, split components
