+++ apps/website/app/components/copy-button/index.ts
@@ -1,0 +1,3 @@
+import CopyButton from "./CopyButton";
+
+export default CopyButton;
