+++ package.json
@@ -14,8 +14,8 @@
     "build:common": "turbo run build --filter=@transi-store/common",
     "build:cli": "turbo run build --filter=@transi-store/cli",
     "start": "turbo run start --filter=@transi-store/website",
-    "download-translations": "yarn transi-store",
-    "upload-translations": "yarn transi-store upload:config",
+    "download-translations": "node -r dotenv/config ./node_modules/.bin/transi-store",
+    "upload-translations": "node -r dotenv/config ./node_modules/.bin/transi-store upload:config",
     "lint:eslint": "eslint .",
     "lint:types": "yarn turbo lint:types",
     "knip": "yarn workspace @transi-store/website react-router typegen && knip",
